% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proyectonuevo-package.R
\docType{package}
\name{proyectonuevo-package}
\alias{proyectonuevo}
\alias{proyectonuevo-package}
\title{proyectonuevo: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/avallecam/proyectonuevo}
  \item Report bugs at \url{https://github.com/avallecam/proyectonuevo/issues}
}

}
\author{
\strong{Maintainer}: Andree Valle-Campos \email{avallecam@gmail.com} (\href{https://orcid.org/0000-0002-7779-481X}{ORCID}) [copyright holder]

}
\keyword{internal}
